requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.10
distro==1.9.0
fake-useragent==1.5.1
fonttools==4.49.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.2.2
magic-filter==1.0.12
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
openai==1.30.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
parse==1.20.1
patsy==0.5.6
pillow==10.2.0
pmdarima==2.0.4
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pyparsing==3.1.1
pyppeteer==2.0.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
pyquery==2.0.0
PySide6==6.7.1
PySide6_Addons==6.7.1
PySide6_Essentials==6.7.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
requests-html==0.10.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.21.0
setuptools==70.0.0
shiboken6==6.7.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
statsmodels==0.14.1
threadpoolctl==3.3.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
w3lib==2.1.2
webdriver-manager==4.0.1
websockets==10.4
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.2
